Fix webMethods Adapter Runtime (ART) connection leaks when using tundra.tn.queue $concurrency > 1 feature to execute queued tasks concurrently with a thread pool.
ART relies on an com.wm.app.b2b.server.invoke.IServiceCompletionListener
object to release an adapter connection, and unfortunately the method
com.wm.app.b2b.server.invoke.ServiceCompletionImpl.serviceCompleted
which notifies listeners of the completion of a service only supports
top-level service completion. So even if a child service is the
service which calls the ART adapter service that opens the
connection, the connection is not released until the top-level
invoking service completes.

When $concurrency > 1, the tundra.tn.queue services use a
permafrost.tundra.server.ServerThreadPoolExecutor object to process
the queued tasks concurrently.

Previously the ServerThreadPoolExecutor was initializing each
ServerThread in the thread pool with a clone of the current
InvokeState that included the current call stack. This meant that
when one of the pooled ServerThread objects invoked a service to
process a queued task, it is invoked with a call stack that already
includes other services (the top-level service and its children).
When the service invoked by the thread completes, the
ServiceCompletionImpl.serviceCompleted(boolean) method checks the
size of the call stack which shows that the completed service was
not the top-level service and therefore it does not notify the ART
service completion listener object to release the connection,
and instead the connection leaks.

This change fixes the ART connection leaks that result from a
tundra.tn.queue service with $concurrency > 1 by not cloning the
call stack when cloning an InvokeState used to initialize a
ServerThread.