Change tundra.tn:deliver to include the delivery result returned by the call to Tundra/tundra.content:deliver in the pipeline when invoking the `$finally` service.
The tundra.content:deliver protocol implementation services can
each return a protocol-specific `$response` and `$response.context`,
which this change makes available for further processing by the
`$finally` service if required.

For example, an HTTP delivery is implemented with the
Tundra/tundra.content.deliver:http service, which returns the HTTP
response body in `$response` and the HTTP return status and
headers in the `$response.context` document.