Add `tundra.tn.document.duplicate:key` service to calculate a unique key for duplicate detection for a given document.
Change `tundra.tn.document.duplicate:content` to return the calculated
key used for duplicate detection, and to store the key on the given
document in the `Unique Key` document attribute if it is defined in
Trading Networks.

Change `tundra.tn.document.duplicate:identity` to return the
calculated key used for duplicate detection, and to store the key on
the given document in the `Unique Key` document attribute if it is
defined in Trading Networks.